Refactor Gemini chat model class
- upgrade to google-genai 1.0.0
- remove special prompt for thinking
