+++ pyproject.toml
@@ -21,7 +21,7 @@ packages = [
 aioconsole = "^0.8.1"
 aiofiles = "^24.1"
 anthropic = "^0.43.0"
-google-genai = "^0.6.0"
+google-genai = "^1.0"
 ipybox = "^0.3.1"
 openai = "^1.59"
 prompt_toolkit = "^3.0"
